[Shop] Refactor Twig components structure
