+++ src/Sylius/Bundle/ShopBundle/Resources/config/app/twig_hooks/account/change_password.yaml
@@ -2,7 +2,7 @@ twig_hooks:
     hooks:
         'sylius_shop.account.change_password':
             form:
-                component: 'sylius_shop:account:change_password:form'
+                component: 'sylius_shop:account:change_password_form'
                 props:
                     form: '@=_context.form'
