+++ src/Sylius/Bundle/ShopBundle/Resources/config/app/twig_hooks/cart/index.yaml
@@ -6,7 +6,7 @@ twig_hooks:
                 props:
                     cart: '@=_context.cart'
             suggested_products:
-                component: 'sylius_shop:common:latest_products'
+                component: 'sylius_shop:product:list'
                 props:
                     limit: 4
                 configuration:
