+++ src/Sylius/Bundle/ShopBundle/Resources/config/app/twig_hooks/homepage/index.yaml
@@ -4,12 +4,14 @@ twig_hooks:
             banner:
                 template: '@SyliusShop/homepage/banner.html.twig'
             latest_deals:
-                component: 'sylius_shop:common:latest_deals'
+                component: 'sylius_shop:product:list'
                 props:
                     limit: 4
+                configuration:
+                    title: 'sylius.homepage.latest_deals'
             new_collection:
                 template: '@SyliusShop/homepage/new_collection.html.twig'
             latest_products:
-                component: 'sylius_shop:common:latest_products'
+                component: 'sylius_shop:product:list'
                 props:
                     limit: 8
