+++ src/Sylius/Bundle/ShopBundle/Twig/Component/Product/PriceComponent.php
@@ -19,16 +19,14 @@ use Sylius\Component\Core\Calculator\ProductVariantPricesCalculatorInterface;
 use Sylius\Component\Core\Model\ProductVariant;
 use Sylius\Component\Currency\Context\CurrencyContextInterface;
 use Sylius\Component\Locale\Context\LocaleContextInterface;
-use Sylius\TwigHooks\LiveComponent\HookableLiveComponentTrait;
-use Symfony\UX\LiveComponent\DefaultActionTrait;
+use Sylius\TwigHooks\Twig\Component\HookableComponentTrait;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 use Symfony\UX\TwigComponent\Attribute\ExposeInTemplate;
 
 #[AsTwigComponent]
 class PriceComponent
 {
-    use DefaultActionTrait;
-    use HookableLiveComponentTrait;
+    use HookableComponentTrait;
 
     public ProductVariant $variant;
