+++ src/Sylius/Bundle/ShopBundle/templates/product/show/page/product_listing/association.html.twig
@@ -3,7 +3,7 @@
     <div {{ sylius_test_html_attribute('product-association', product_association.type.name) }}>
         <div class="products-grid">
             {% for product in associated_products %}
-                <div>{{ component('sylius_shop:common:product_card', { product: product }) }}</div>
+                <div>{{ component('sylius_shop:product:card', { product: product }) }}</div>
             {% endfor %}
         </div>
     </div>
