+++ src/Sylius/Bundle/ShopBundle/templates/shared/product/horizontal_list.html.twig
@@ -1,5 +1,5 @@
 <div class="products-grid">
     {% for product in products %}
-        <div>{{ component('sylius_shop:common:product_card', { product: product }) }}</div>
+        <div>{{ component('sylius_shop:product:card', { product: product }) }}</div>
     {% endfor %}
 </div>
